+++ Speaker.h
@@ -7,25 +7,25 @@ class Speaker
     public:
         Speaker(byte);
         Speaker(byte, unsigned short);
-        Speaker(byte, unsigned short, unsigned int);
+        Speaker(byte, unsigned short, unsigned long);
         byte getPinNumber();
         bool getCurrentState();
         unsigned short getToneInHertz();
-        unsigned int getDurationInMilliseconds();
+        unsigned long getDurationInMilliseconds();
         void turnOn();
         void turnOn(unsigned short);
         void turnOff();
         void blink();
         void blink(unsigned short);
-        void blink(unsigned short, unsigned int);
+        void blink(unsigned short, unsigned long);
 
     private:
         byte _pinNumber;
         bool _currentState;
         unsigned short _toneInHertz;
-        unsigned int _durationInMilliseconds;
+        unsigned long _durationInMilliseconds;
         unsigned long _timestampInMilliseconds;
-        bool _isTheRightTimeToChangeState(unsigned long, unsigned int);
+        bool _isTheRightTimeToChangeState(unsigned long, unsigned long);
         void _changeCurrentState();
         void _decideWetherToTurnOnOrOff();
 };
